pyproject.toml
[project]
name = "rss-intelligence"
version = "0.1.0"
description = "Local-first RSS intelligence CLI"
readme = "README.md"
authors = [{ name = "RSS Intel" }]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "requests>=2.32",
    "feedparser>=6.0",
    "pydantic>=2.7",
    "PyYAML>=6.0",
    "rich>=13.7"
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
news = "news.cli:app"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
